+++ day2/day2_test.go
@@ -7,9 +7,9 @@ import (
 )
 
 func Test_solvePart1(t *testing.T) {
-	assert.Equal(t, 5434663, solvePart1(intcode.NewIntcode(day2_program...)))
+	assert.Equal(t, int64(5434663), solvePart1(intcode.NewIntcode(day2_program...)))
 }
 
 func Test_solvePart2(t *testing.T) {
-	assert.Equal(t, 4559, solvePart2(intcode.NewIntcode(day2_program...)))
+	assert.Equal(t, int64(4559), solvePart2(intcode.NewIntcode(day2_program...)))
 }
