+++ day5/day5_test.go
@@ -6,9 +6,9 @@ import (
 )
 
 func Test_solvePart1(t *testing.T) {
-	assert.Equal(t, 8332629, solvePart1())
+	assert.Equal(t, int64(8332629), solvePart1())
 }
 
 func Test_solvePart2(t *testing.T) {
-	assert.Equal(t, 8805067, solvePart2())
+	assert.Equal(t, int64(8805067), solvePart2())
 }
